tightened up identifcations with new gameserver id

+++ bin/cleanup.sh
@@ -2,13 +2,18 @@
 
 whoami=$(whoami | tr -d '[:space:]')
 
+if [[ "$whoami" != "sembiance" ]]; then
+	echo "Error: This script must be run as 'sembiance', not '$whoami'." >&2
+	exit 1
+fi
+
 pkill -f dexserver
 kill "$(cat /mnt/dexvert/daemonize_dexserver.pid)"
 kill "$(pidof -x dexserver)"
 sudo killall -q -9 winedevice.exe services.exe explorer.exe plugplay.exe svchost.exe rpcss.exe sf dosbox
 killall Xvfb
 killall -9 at-spi-bus-launcher helpdeco gimp-org-file-pnm identify convert 86Box timidity gimp x11vnc uadecore ffmpeg
-killall wineserver python3 sf java uniconvertor oosplash soffice.bin texi2dvi DirectorCastRipper.exe pdfetex pdflatex tail
+killall wineserver python3 sf java uniconvertor oosplash soffice.bin texi2dvi DirectorCastRipper.exe pdfetex pdflatex tail arj
 
 if [[ "$(hostname)" == dexdrone* ]]; then
 	kill $(pgrep -f /mnt/compendium/DevLab/dexvert/bin/startDexserver)
